# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Twinship API
  description: |
    RESTful API for Twinship - A specialized mobile platform for twins to strengthen their bond and connect through chat, games, assessments, and stories.
    
    This API provides endpoints for:
    - User authentication and profile management
    - Twin pairing and invitation system
    - Real-time messaging and communication
    - Assessment and research data collection
    - Premium subscription management
    - Research participation and telemetry
    
  version: 1.0.0
  contact:
    name: Twinship API Support
    email: api@twinship.app
  license:
    name: Private
servers:
  - url: https://api.twinship.app/v1
    description: Production server
  - url: https://staging-api.twinship.app/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user
      description: Create a new user account with email and password
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - birthDate
                - twinType
                - accentColor
              properties:
                email:
                  type: string
                  format: email
                  example: "jordan@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "securePassword123!"
                name:
                  type: string
                  minLength: 1
                  example: "Jordan Smith"
                age:
                  type: integer
                  minimum: 1
                  maximum: 120
                  example: 25
                gender:
                  type: string
                  example: "female"
                sexualOrientation:
                  type: string
                  example: "heterosexual"
                showSexualOrientation:
                  type: boolean
                  default: false
                birthDate:
                  type: string
                  format: date
                  example: "1998-06-15"
                placeOfBirth:
                  type: string
                  example: "San Francisco, CA"
                timeOfBirth:
                  type: string
                  example: "14:30"
                twinType:
                  $ref: '#/components/schemas/TwinType'
                otherTwinTypeDescription:
                  type: string
                  description: "Required if twinType is 'other'"
                twinDeceased:
                  type: boolean
                  default: false
                accentColor:
                  $ref: '#/components/schemas/ThemeColor'
                profilePicture:
                  type: string
                  format: uri
                  description: "URL to profile picture"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    $ref: '#/components/schemas/UserProfile'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login user
      description: Authenticate user with email and password
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                rememberMe:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/UserProfile'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/refresh:
    post:
      summary: Refresh access token
      description: Generate new access token using refresh token
      tags:
        - Authentication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    TwinType:
      type: string
      enum: [identical, fraternal, other]
      example: "identical"
    ThemeColor:
      type: string
      enum: [neon-pink, neon-blue, neon-green, neon-yellow, neon-purple, neon-orange, neon-cyan, neon-red]
      example: "neon-purple"
    UserProfile:
      type: object
      properties:
        id:
          type: string
          example: "user_123456789"
        name:
          type: string
          example: "Jordan Smith"
        age:
          type: integer
          example: 25
        gender:
          type: string
          example: "female"
        sexualOrientation:
          type: string
          example: "heterosexual"
        showSexualOrientation:
          type: boolean
          example: false
        twinType:
          $ref: '#/components/schemas/TwinType'
        otherTwinTypeDescription:
          type: string
        twinDeceased:
          type: boolean
          example: false
        birthDate:
          type: string
          format: date
          example: "1998-06-15"
        zodiacSign:
          type: string
          example: "Gemini"
        placeOfBirth:
          type: string
          example: "San Francisco, CA"
        timeOfBirth:
          type: string
          example: "14:30"
        profilePicture:
          type: string
          format: uri
        accentColor:
          $ref: '#/components/schemas/ThemeColor'
        isConnected:
          type: boolean
          example: true
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
          description: "JWT access token (expires in 1 hour)"
        refreshToken:
          type: string
          description: "Refresh token (expires in 30 days)"
        expiresIn:
          type: integer
          description: "Access token expiration time in seconds"
          example: 3600
        tokenType:
          type: string
          example: "Bearer"
    
    # Invitation Types
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
              example: { "field": "email", "reason": "invalid_format" }
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...